requirements.txt
colorama==0.4.6
customtkinter==5.1.2
discord.py==2.1.1
requests==2.28.2
